requirements.txt
aiogram==3.13.1
SQLAlchemy[asyncio]==2.0.36
aiosqlite==0.20.0
openpyxl==3.1.5
python-dotenv==1.0.1
pytest>=7.0.0
pytest-asyncio>=0.21.0

